+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 All notable changes to the "Ternary ⇄ If-Else Converter" extension will be documented in this file.
 
-## Added
+## [Added]
 - v0.0.1 Ternary ⇄ If-Else Converter
+- v0.0.2 Toggle Button + Granular level conversion
+    - Added Toggle Button To Enable And Disable Hover View
+    - Added Granular level conversion
+    - Complexity Support Increase
+    - Updated Error Handling
 
 ## [Unreleased]